+++ mobile/package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "@expo-google-fonts/nunito": "^0.1.0",
+    "@react-native-community/async-storage": "^1.12.1",
     "@react-native-community/masked-view": "0.1.10",
     "@react-navigation/native": "^5.7.6",
     "@react-navigation/stack": "^5.9.3",
